+++ docs/changelog.rst
@@ -49,6 +49,8 @@ Detailed list of changes
 - Wayland: Fix for a bug that could cause kitty to become non-responsive when
   using multiple OS windows in a single instance on some compositors (:iss:`5495`)
 
+- Wayland: Fix for a bug preventing kitty from starting on Hyprland when using a non-unit scale (:iss:`5467`)
+
 0.26.2 [2022-09-05]
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~